+++ spacer/config.py
@@ -252,8 +252,11 @@ STORAGE_TYPES = [
 MAX_IMAGE_PIXELS = get_config_value('MAX_IMAGE_PIXELS', default=10000*10000)
 MAX_POINTS_PER_IMAGE = get_config_value('MAX_POINTS_PER_IMAGE', default=1000)
 
-# The train_classifier task requires as least this many images.
-MIN_TRAINIMAGES = get_config_value('MIN_TRAINIMAGES', default=10)
+# Size of training batches. This number of features must be able to fit
+# in memory. Raising this allows the reference set to be larger,
+# which can improve calibration results.
+TRAINING_BATCH_LABEL_COUNT = get_config_value(
+    'TRAINING_BATCH_LABEL_COUNT', default=5000)
 
 # Check access to select which tests to run.
 HAS_CAFFE = importlib.util.find_spec("caffe") is not None
@@ -288,7 +291,6 @@ CONFIGURABLE_VARS = [
     'LOG_LEVEL',
     'MAX_IMAGE_PIXELS',
     'MAX_POINTS_PER_IMAGE',
-    'MIN_TRAINIMAGES',
 ]