+++ spacer/data_classes.py
@@ -72,7 +72,15 @@ class ImageLabels(DataClass):
                  # Data maps a feature key (or file path) to a List of
                  # (row, col, label).
                  data: dict[str, list[tuple[int, int, int]]]):
-        self.data = data
+        self._data = data
+
+        self.label_count = sum([len(labels) for labels in data.values()])
+
+        self.classes_set = set()
+        for single_image_labels in data.values():
+            single_image_classes = set(
+                label for (row, col, label) in single_image_labels)
+            self.classes_set |= single_image_classes
 
     @classmethod
     def example(cls):
@@ -83,30 +91,48 @@ class ImageLabels(DataClass):
             'img4.features': [(100, 202, 3), (101, 200, 3)],
         })
 
+    def serialize(self) -> dict:
+        """Only need the `_data` field; the other fields can be recomputed."""
+        return self._data
+
     @classmethod
     def deserialize(cls, data: dict) -> 'ImageLabels':
-        """ Custom deserializer required to convert back to tuples. """
-        return ImageLabels(
-            data={key: [tuple(entry) for entry in value] for
-                  key, value in data['data'].items()})
+        """Custom deserializer required to convert back to tuples."""
+        return ImageLabels({
+            key: [tuple(entry) for entry in value] for
+            key, value in data.items()
+        })
 
     @property
     def image_keys(self):
-        return list(self.data.keys())
+        return list(self._data.keys())
 
-    @property
-    def samples_per_image(self):
-        return len(next(iter(self.data.values())))
-
-    def unique_classes(self, key_list: list[str]) -> set[int]:
-        """ Returns the set of all unique classes in the key_list subset. """
-        labels = set()
-        for im_key in key_list:
-            labels |= {label for (row, col, label) in self.data[im_key]}
-        return labels
+    def filter_classes(self, accepted_classes) -> 'ImageLabels':
+        """
+        Make a new instance by filtering out labels not included in
+        the specified classes.
+        """
+        data = {}
+        for image_key in self.image_keys:
+            this_image_labels = [
+                (row, column, label)
+                for row, column, label in self._data[image_key]
+                if label in accepted_classes
+            ]
+            # Only include an image if it has any labels remaining
+            # after filtering.
+            if len(this_image_labels) > 0:
+                data[image_key] = this_image_labels
+        return ImageLabels(data)
 
     def __len__(self):
-        return len(self.data)
+        return len(self._data)
+
+    def __getitem__(self, item):
+        return self._data[item]
+
+    def __contains__(self, item):
+        return item in self._data
 
 
 class PointFeatures(DataClass):