+++ spacer/train_utils.py
@@ -5,7 +5,9 @@ Utility methods for training classifiers.
 from __future__ import annotations
 import random
 import string
+from collections.abc import Generator
 from logging import getLogger
+from typing import List, Tuple
 
 import numpy as np
 from sklearn.calibration import CalibratedClassifierCV
@@ -20,69 +22,58 @@ from spacer.messages import DataLocation
 logger = getLogger(__name__)
 
 
-def train(labels: ImageLabels,
+# Implicit type alias; revisit in Python 3.10
+# https://peps.python.org/pep-0613/
+FeatureLabelPair = Tuple[np.ndarray, int]
+FeatureLabelBatch = Tuple[List[np.ndarray], List[int]]
+
+
+def train(train_labels: ImageLabels,
+          ref_labels: ImageLabels,
           feature_loc: DataLocation,
           nbr_epochs: int,
           clf_type: str) -> tuple[CalibratedClassifierCV, list[float]]:
 
-    if len(labels) < config.MIN_TRAINIMAGES:
-        raise ValueError('Not enough training samples.')
+    logger.debug(
+        f"Data sets:"
+        f" Train = {len(train_labels)} images,"
+        f" {train_labels.label_count} labels;"
+        f" Ref = {len(ref_labels)} images,"
+        f" {ref_labels.label_count} labels")
+    logger.debug(
+        f"Mini-batch size: {config.TRAINING_BATCH_LABEL_COUNT} labels")
 
-    # Calculate max nbr images to keep in memory (for 5000 samples total).
-    max_imgs_in_memory = 5000 // labels.samples_per_image
-
-    # Make train and ref split.
-    # Reference set is here a hold-out part of the train-data portion.
-    # Purpose of reference set is to
-    # 1) know accuracy per epoch
-    # 2) calibrate classifier output scores.
-    # We call it "reference" to disambiguate from the validation set.
-    ref_set = labels.image_keys[::10]
-    np.random.shuffle(ref_set)
-    ref_set = ref_set[:max_imgs_in_memory]  # Enforce memory limit.
-    train_set = list(set(labels.image_keys) - set(ref_set))
-    logger.debug("Trainset: {}, valset: {} images".
-                 format(len(train_set), len(ref_set)))
-
-    # Figure out # images per batch and batches per epoch.
-    images_per_batch, batches_per_epoch = \
-        calc_batch_size(max_imgs_in_memory, len(train_set))
-    logger.debug("Using {} images per mini-batch and {} mini-batches per "
-                 "epoch".format(images_per_batch, batches_per_epoch))
-
-    # Identify classes common to both train and val.
-    # This will be our labelset for the training.
-    trainclasses = labels.unique_classes(train_set)
-    refclasses = labels.unique_classes(ref_set)
-    classes = list(trainclasses.intersection(refclasses))
-    logger.debug("Trainset: {}, valset: {}, common: {} labels".format(
-        len(trainclasses), len(refclasses), len(classes)))
-    if len(classes) == 1:
-        raise ValueError('Not enough classes to do training (only 1)')
+    # train_labels and ref_labels should already be trimmed down to the
+    # classes common to both, so the classes_set from either one should be the
+    # set used for training.
+    classes_list = list(ref_labels.classes_set)
 
     # Load reference data (must hold in memory for the calibration)
     with config.log_entry_and_exit("loading of reference data"):
-        refx, refy = load_batch_data(labels, ref_set, classes, feature_loc)
+        refx, refy = load_batch_data(ref_labels, feature_loc)
 
     # Initialize classifier and ref set accuracy list
     with config.log_entry_and_exit("training using " + clf_type):
         if clf_type == 'MLP':
-            if len(train_set) * labels.samples_per_image >= 50000:
+            if train_labels.label_count >= 50000:
                 hls, lr = (200, 100), 1e-4
             else:
                 hls, lr = (100,), 1e-3
             clf = MLPClassifier(hidden_layer_sizes=hls, learning_rate_init=lr)
         else:
             clf = SGDClassifier(loss='log_loss', average=True, random_state=0)
+
         ref_acc = []
+
         for epoch in range(nbr_epochs):
-            np.random.shuffle(train_set)
-            mini_batches = chunkify(train_set, batches_per_epoch)
-            for mb in mini_batches:
-                x, y = load_batch_data(labels, mb, classes, feature_loc)
-                clf.partial_fit(x, y, classes=classes)
+            for x, y in load_data_as_mini_batches(
+                labels=train_labels, feature_loc=feature_loc,
+                random_state=epoch,
+            ):
+                clf.partial_fit(x, y, classes=classes_list)
+
             ref_acc.append(calc_acc(refy, clf.predict(refx)))
-            logger.debug("Epoch {}, acc: {}".format(epoch, ref_acc[-1]))
+            logger.debug(f"Epoch {epoch}, acc: {ref_acc[-1]}")
 
     with config.log_entry_and_exit("calibration"):
         clf_calibrated = CalibratedClassifierCV(clf, cv="prefit")
@@ -93,114 +84,153 @@ def train(labels: ImageLabels,
 
 def evaluate_classifier(clf: CalibratedClassifierCV,
                         labels: ImageLabels,
-                        classes: list[int],
                         feature_loc: DataLocation) -> tuple[list, list, list]:
     """ Evaluates classifier on data """
     scores, gts, ests = [], [], []
-    for imkey in labels.image_keys:
-        x, y = load_image_data(labels, imkey, classes, feature_loc)
-        if len(x) > 0:
-            scores.extend(clf.predict_proba(x).max(axis=1).tolist())
-            ests.extend(clf.predict(x))
-            gts.extend(y)
 
-    if len(gts) == 0:
-        raise ValueError('Not enough data in validation set.')
+    for image_key in labels.image_keys:
+
+        feature_loc.key = image_key
+        image_labels_data = labels[image_key]
+
+        pairs = list(
+            load_image_data(image_labels_data, feature_loc))
+
+        # List of pairs -> pair of lists
+        x, y = zip(*pairs)
+
+        scores.extend(clf.predict_proba(x).max(axis=1).tolist())
+        ests.extend(clf.predict(x))
+        gts.extend(y)
+
+    assert len(gts) > 0, (
+        "The validation set should have been checked for emptiness during"
+        " label preprocessing.")
 
     return gts, ests, scores
 
 
-def chunkify(lst: list,
-             nbr_chunks: int) -> list:
-    return [lst[i::nbr_chunks] for i in range(nbr_chunks)]
-
-
-def calc_batch_size(max_imgs_in_memory: int,
-                    train_set_size: int) -> tuple[int, int]:
-    images_per_batch = min(max_imgs_in_memory, train_set_size)
-    batches_per_epoch = int(np.ceil(train_set_size / images_per_batch))
-    return images_per_batch, batches_per_epoch
-
-
-def load_image_data(labels: ImageLabels,
-                    imkey: str,
-                    classes: list[int],
+def load_image_data(labels_data: list[tuple[int, int, int]],
                     feature_loc: DataLocation) \
-        -> tuple[list[list[float]], list[int]]:
+        -> Generator[FeatureLabelPair, None, None]:
     """
-    Loads features and labels for the specified image,
-    and builds element-matching lists for use with methods such as
-    CalibratedClassifierCV.fit().
+    Loads a feature vector and labels of a single image, and generates
+    element-matching pairs.
     """
+    # Load features.
+    features = ImageFeatures.load(feature_loc)
 
-    # Load features for this image.
-    feature_loc.key = imkey
-    image_features = ImageFeatures.load(feature_loc)
+    return match_features_and_labels(features, labels_data, feature_loc.key)
 
-    # Load row, col, labels for this image.
-    image_labels = labels.data[imkey]
+
+def load_batch_data(labels: ImageLabels,
+                    feature_loc: DataLocation) \
+        -> FeatureLabelBatch:
+    """
+    Loads features and labels, and builds element-matching lists
+    for use with methods such as CalibratedClassifierCV.fit().
+    """
+    batch = []
+
+    for image_key in labels.image_keys:
+
+        feature_loc.key = image_key
+        image_labels_data = labels[image_key]
+
+        batch.extend(
+            load_image_data(image_labels_data, feature_loc))
+
+    assert len(batch) > 0, (
+        "We only ever expect labels to be a non-empty ref set.")
+
+    # List of pairs -> pair of lists
+    x, y = zip(*batch)
+    return x, y
+
+
+def load_data_as_mini_batches(labels: ImageLabels,
+                              feature_loc: DataLocation,
+                              # For seeding the randomizer to get repeatable
+                              # results.
+                              random_state: int) \
+        -> Generator[FeatureLabelBatch, None, None]:
+    """
+    Loads features and labels, and generates batches of
+    element-matching pairs.
+
+    Since this is a generator, it does not have to load all feature
+    vectors in memory at the same time; only as many as will fit in
+    a batch.
+    Note that a single image's features may straddle multiple batches.
+    """
+    image_keys = labels.image_keys
+
+    # Shuffle the order of images.
+    np.random.seed(random_state)
+    np.random.shuffle(image_keys)
+
+    current_batch = []
+
+    for image_key in image_keys:
+
+        feature_loc.key = image_key
+        image_labels_data = labels[image_key]
+
+        for point_feature, label in load_image_data(
+                image_labels_data, feature_loc):
+
+            current_batch.append((point_feature, label))
+
+            if len(current_batch) >= config.TRAINING_BATCH_LABEL_COUNT:
+                # List of pairs -> pair of lists
+                x, y = zip(*current_batch)
+                yield x, y
+                current_batch = []
+
+    if len(current_batch) > 0:
+        # Last batch
+        x, y = zip(*current_batch)
+        yield x, y
+
+
+def match_features_and_labels(features: ImageFeatures,
+                              labels_data: list[tuple[int, int, int]],
+                              image_key: str) \
+        -> Generator[FeatureLabelPair, None, None]:
 
     # Sanity check
-    if image_features.valid_rowcol:
+    if features.valid_rowcol:
         # With new data structure just check that the sets of row, col
         # given by the labels are available in the features.
         rc_features_set = set([(pf.row, pf.col) for pf in
-                               image_features.point_features])
-        rc_labels_set = set([(row, col) for (row, col, _) in image_labels])
+                               features.point_features])
+        rc_labels_set = set([(row, col) for (row, col, _) in labels_data])
 
         if not rc_labels_set.issubset(rc_features_set):
             difference_set = rc_labels_set.difference(rc_features_set)
             example_rc = next(iter(difference_set))
             raise RowColumnMismatchError(
-                f"{imkey}: The labels' row-column positions don't match those"
-                f" of the feature vector (example: {example_rc}).")
+                f"{image_key}: The labels' row-column positions don't match"
+                f" those of the feature vector (example: {example_rc}).")
     else:
         # With legacy data structure check that length is the same.
-        label_count = len(image_labels)
-        feature_count = len(image_features.point_features)
+        label_count = len(labels_data)
+        feature_count = len(features.point_features)
 
         if not label_count == feature_count:
             raise RowColumnMismatchError(
-                f"{imkey}: The number of labels ({label_count}) doesn't match"
-                f" the number of extracted features ({feature_count}).")
+                f"{image_key}: The number of labels ({label_count}) doesn't"
+                f" match the number of extracted features ({feature_count}).")
 
-    x, y = [], []
-    if image_features.valid_rowcol:
-        for row, col, label in image_labels:
-            if label not in classes:
-                # Remove samples for which the label is not in classes.
-                continue
-            x.append(image_features[(row, col)])
-            y.append(label)
-
+    if features.valid_rowcol:
+        for row, col, label in labels_data:
+            yield features[(row, col)], label
     else:
         # For legacy features, we didn't store the row, col information.
         # Instead rely on ordering.
-        for (_, _, label), point_feature in zip(image_labels,
-                                                image_features.point_features):
-            if label not in classes:
-                continue
-            x.append(point_feature.data)
-            y.append(label)
-
-    return x, y
-
-
-def load_batch_data(labels: ImageLabels,
-                    imkeylist: list[str],
-                    classes: list[int],
-                    feature_loc: DataLocation) \
-        -> tuple[list[list[float]], list[int]]:
-    """
-    Builds element-matching lists of features and labels
-    for multiple images.
-    """
-    x, y = [], []
-    for imkey in imkeylist:
-        x_, y_ = load_image_data(labels, imkey, classes, feature_loc)
-        x.extend(x_)
-        y.extend(y_)
-    return x, y
+        for (_, _, label), point_feature in zip(labels_data,
+                                                features.point_features):
+            yield point_feature.data, label
 
 
 def calc_acc(gt: list[int], est: list[int]) -> float:
@@ -233,7 +263,7 @@ def make_random_data(im_count: int,
     Utility method for testing that generates an ImageLabels instance
     complete with stored ImageFeatures.
     """
-    labels = ImageLabels(data={})
+    data = {}
     for _ in range(im_count):
 
         # Generate random features (using labels to draw from a Gaussian).
@@ -250,8 +280,8 @@ def make_random_data(im_count: int,
         # Store
         feature_loc.key = imkey
         feats.store(feature_loc)
-        labels.data[imkey] = [
+        data[imkey] = [
             (pf.row, pf.col, pl) for pf, pl in
             zip(feats.point_features, point_labels)
         ]
-    return labels
+    return ImageLabels(data)